feat(forgot-password): implement forgot password UI and view model with email validation
