+++ app/src/main/java/com/nullpointer/devs/drivers/presentation/ui/auth/login/LoginComponent.kt
@@ -1,4 +1,4 @@
-package com.nullpointer.devs.drivers.presentation.ui.register
+package com.nullpointer.devs.drivers.presentation.ui.auth.login
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -28,43 +28,43 @@ import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.SavedStateHandle
 import com.nullpointer.devs.drivers.R
 import com.nullpointer.devs.drivers.presentation.ui.components.PasswordFieldComponent
 import com.nullpointer.devs.drivers.presentation.ui.components.TextFieldComponent
-import com.nullpointer.devs.drivers.presentation.ui.login.state.InputState
-import com.nullpointer.devs.drivers.presentation.ui.login.state.PasswordState
+import com.nullpointer.devs.drivers.presentation.ui.auth.login.state.InputState
+import com.nullpointer.devs.drivers.presentation.ui.auth.login.state.PasswordState
 
 
 @Composable
-fun RegisterComponent(
-    registerViewModel: RegisterViewModel = hiltViewModel()
+fun LoginComponent(
+    viewModel: LoginViewModel = hiltViewModel()
 ) {
 
-    RegisterComponent(
-        emailInputState = registerViewModel.emailInputState,
-        passwordState = registerViewModel.passwordInputState,
-        lastNameState = registerViewModel.lastNameInputState,
-        nameState = registerViewModel.nameInputState,
-        birthDateState = registerViewModel.birthdayInputState,
+    LoginComponent(
+        modifier = Modifier.fillMaxSize(),
+        emailInputState = viewModel.emailInputState,
+        passwordState = viewModel.passwordInputState,
         validateForm = {
-
-        },
-        registerAction = {
-
+            viewModel.validateFields()?.let {
+                viewModel.login(it)
+            }
         }
     )
 }
 
+
+
+
+
 @Composable
-private fun RegisterComponent(
+private fun LoginComponent(
     emailInputState: InputState,
     passwordState: PasswordState,
-    lastNameState: InputState,
-    nameState: InputState,
-    birthDateState: InputState,
     modifier: Modifier = Modifier,
     validateForm: () -> Unit = {},
-    registerAction: () -> Unit = {}
+    registerAction: () -> Unit = {},
+    forgotPasswordAction: () -> Unit = {}
 ) {
     Scaffold(
         modifier = modifier
@@ -97,25 +97,11 @@ private fun RegisterComponent(
                     horizontalAlignment = Alignment.CenterHorizontally,
                 ) {
                     Text(
-                        text = stringResource(R.string.register_text),
+                        text = stringResource(R.string.login_text),
                         style = MaterialTheme.typography.headlineSmall,
                         color = Color.Black
                     )
 
-                    TextFieldComponent(
-                        state = nameState,
-                        modifier = Modifier.fillMaxWidth()
-                    )
-
-                    TextFieldComponent(
-                        state = lastNameState,
-                        modifier = Modifier.fillMaxWidth()
-                    )
-
-                    TextFieldComponent(
-                        state = birthDateState,
-                        modifier = Modifier.fillMaxWidth()
-                    )
 
                     TextFieldComponent(
                         state = emailInputState,
@@ -127,67 +113,81 @@ private fun RegisterComponent(
                         modifier = Modifier.fillMaxWidth()
                     )
 
+                    Column(
+                        modifier = Modifier.fillMaxWidth(),
+                        verticalArrangement = Arrangement.spacedBy(8.dp),
+                        horizontalAlignment = Alignment.CenterHorizontally,
+                    ){
+                        ExtendedFloatingActionButton(
+                            onClick = validateForm,
+                            modifier = Modifier.width(200.dp)
+                        ) {
+                            Text(text = stringResource(R.string.login_text))
+                        }
 
-                    ExtendedFloatingActionButton(
-                        onClick = validateForm,
-                        modifier = Modifier.width(200.dp)
-                    ) {
-                        Text(text = stringResource(R.string.register_text_button))
+                        TextButton(
+                            onClick = forgotPasswordAction,
+                            modifier = Modifier.padding(start = 4.dp)
+                        ) {
+                            Text(
+                                text = stringResource(R.string.recover),
+                                style = MaterialTheme.typography.bodySmall,
+                                color = MaterialTheme.colorScheme.primary
+                            )
+                        }
                     }
 
+
+
+                    Row(
+                        modifier = Modifier.fillMaxWidth(),
+                        horizontalArrangement = Arrangement.Center,
+                        verticalAlignment = Alignment.CenterVertically
+                    ) {
+                        Text(
+                            text = stringResource(R.string.no_has_account),
+                            style = MaterialTheme.typography.bodySmall,
+                            color = Color.Black
+                        )
+                        TextButton(
+                            onClick = registerAction,
+                            modifier = Modifier.padding(start = 4.dp)
+                        ) {
+                            Text(
+                                text = stringResource(R.string.sign_up),
+                                style = MaterialTheme.typography.bodySmall,
+                                color = MaterialTheme.colorScheme.primary
+                            )
+                        }
+                    }
                 }
             }
         }
     }
 }
-
 @Preview(
     showBackground = true,
     showSystemUi = true,
     device = Devices.PIXEL_4_XL
 )
 @Composable
-private fun RegisterComponentPreview() {
-    RegisterComponent(
+private fun LoginComponentPreview() {
+    LoginComponent(
         emailInputState = InputState(
             currentValue = "value",
             label = null,
             hint = null,
             validators = emptyList(),
-            savedStateHandle = androidx.lifecycle.SavedStateHandle(),
-            key = ""
+            key = "",
+            savedStateHandle = SavedStateHandle()
         ),
         passwordState = PasswordState(
             currentValue = "value",
             label = null,
             hint = null,
             validators = emptyList(),
-            savedStateHandle = androidx.lifecycle.SavedStateHandle(),
-            key = ""
-        ),
-        lastNameState = InputState(
-            currentValue = "value",
-            label = null,
-            hint = null,
-            validators = emptyList(),
-            savedStateHandle = androidx.lifecycle.SavedStateHandle(),
-            key = ""
-        ),
-        nameState = InputState(
-            currentValue = "value",
-            label = null,
-            hint = null,
-            validators = emptyList(),
-            savedStateHandle = androidx.lifecycle.SavedStateHandle(),
-            key = ""
-        ),
-        birthDateState = InputState(
-            currentValue = "value",
-            label = null,
-            hint = null,
-            validators = emptyList(),
-            savedStateHandle = androidx.lifecycle.SavedStateHandle(),
-            key = ""
+            key = "",
+            savedStateHandle = SavedStateHandle()
         )
     )
 }
@@ -198,47 +198,23 @@ private fun RegisterComponentPreview() {
     device = Devices.TABLET
 )
 @Composable
-private fun RegisterComponentPreviewTablet() {
-    RegisterComponent(
+private fun LoginComponentPreviewTablet() {
+    LoginComponent(
         emailInputState = InputState(
             currentValue = "value",
             label = null,
             hint = null,
             validators = emptyList(),
-            savedStateHandle = androidx.lifecycle.SavedStateHandle(),
-            key = ""
+            key = "",
+            savedStateHandle = SavedStateHandle()
         ),
         passwordState = PasswordState(
             currentValue = "value",
             label = null,
             hint = null,
             validators = emptyList(),
-            savedStateHandle = androidx.lifecycle.SavedStateHandle(),
-            key = ""
-        ),
-        lastNameState = InputState(
-            currentValue = "value",
-            label = null,
-            hint = null,
-            validators = emptyList(),
-            savedStateHandle = androidx.lifecycle.SavedStateHandle(),
-            key = ""
-        ),
-        nameState = InputState(
-            currentValue = "value",
-            label = null,
-            hint = null,
-            validators = emptyList(),
-            savedStateHandle = androidx.lifecycle.SavedStateHandle(),
-            key = ""
-        ),
-        birthDateState = InputState(
-            currentValue = "value",
-            label = null,
-            hint = null,
-            validators = emptyList(),
-            savedStateHandle = androidx.lifecycle.SavedStateHandle(),
-            key = ""
+            key = "",
+            savedStateHandle = SavedStateHandle()
         )
     )
 }
